api/requirements.txt
fastapi 
langchain-google-genai 
langchain
fastapi 
uvicorn
python-dotenv